+++ Sources/Barktor/MenuBarController.swift
@@ -40,6 +40,7 @@ final class MenuBarController {
         self.onShowOnboarding = onShowOnboarding
         self.onQuit = onQuit
         self.statusItem = NSStatusBar.system.statusItem(withLength: NSStatusItem.variableLength)
+        statusItem.autosaveName = "com.naktor.barktor.statusItem"
 
         configureButton()
         rebuildMenu()